+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Changed
+- Removed duplicates from scopes generated strings.
+
+## [unreleased]
+
+### Changed
 - Stop adding ?schema=openid to userinfo endpoint URL. #449
 
 ## [1.0.1] - 2024-09-13