+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Stop adding ?schema=openid to userinfo endpoint URL. #449
 
+### Fixed
+- Fix missing `iat` verification when verifying JWT #476
+
 ## [1.0.1] - 2024-09-13
 
 ### Fixed