+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stop adding ?schema=openid to userinfo endpoint URL. #449
 
 ### Fixed
+- Fix missing `iat` verification when verifying JWT #476
 - Check existence of subject when verifying JWT #474
 - exp verification when verifying Logout Token claims #482