+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support to change the `leeway` time for JWT verification using `setLeeway` #483
 
 ### Changed
+- Added support for `X-Forwarded-Host` & `X-Forwarded-Prefix` headers
 - Stop adding ?schema=openid to userinfo endpoint URL. #449
 
 ### Fixed